app_notificator/app/requirements.txt
Flask
loguru
pymysql
requests
python-dotenv
